requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
facenet-pytorch==2.5.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
python-multipart==0.0.9
